## 4.2.0.0
* Add helloworld example plugin to distribution.
* ATAK CIV plugins are compatible with all variants of ATAK.
* 3D Billboard Feature.  Marker image attachments can appear as 3D Billboards when navigating routes.
* Point Dropper - 3D Vehicle Icons improvement. Higher contrast outlines are rendered around the icons in the point dropper selector.
* Edit Route Checkpoint Elevation. Elevation of individual route checkpoints may now be edited.
* Toggle Sensor Field of View (FOV). The sensor FOV can now be toggled from the sensor radial menu.
* Additional Metrics Collection. Information on screen touches is now included in metrics logging, when enabled.
* Added DTED Data Manager.
* Added GeoPDF support.
* Upgrade to GDAL 2.4.4.
* ASSIMP now included as tarball.
* Upgraded to `atak-gradle-takdev` 2.0.0.
* Various bug fixes

+++ takengine/mapengine/sdk/src/renderer/GLMatrix.cpp
@@ -107,10 +107,10 @@ namespace atakmap
         {
 #if 1
             const auto theta = static_cast<float>(degrees * M_PI / 180.0f);
-            const float sinTheta = sin(theta);
-            const float cosTheta = cos(theta);
+            const float sinTheta = static_cast<float>(sin(theta));
+            const float cosTheta = static_cast<float>(cos(theta));
 
-            const float mag = sqrt((axisX*axisX) + (axisY*axisY) + (axisZ*axisZ));
+            const float mag = static_cast<float>(sqrt((axisX*axisX) + (axisY*axisY) + (axisZ*axisZ)));
 
             const float nX = axisX / mag;
             const float nY = axisY / mag;
